Reorganise GUI implementation files for strategic form
