+++ src/gui/nfgtable.h
@@ -27,9 +27,53 @@
 
 #include "games/stratpure.h"
 
+// Forward-declared rather than requiring <wx/print.h>: TableWidget::GetPrintout() only
+// ever needs the pointer type here, and not every .cc file including this header (e.g.
+// nfgheadergrid.cc) has a reason to pull in the printing headers itself.
+class wxPrintout;
+
 namespace Gambit::GUI {
 class GameDocument;
 class NfgPanel;
+
+//!
+//! This class handles some common overriding of wxGrid behavior common to the grids used
+//! in the strategic game display (nfgtable.cc's TableGridBase, nfgheadergrid.cc's
+//! RowPlayerGrid/ColPlayerGrid, nfgpayoffgrid.cc's PayoffGrid) -- shared here, rather than
+//! file-local to one of them, since it's a base class used across all three .cc files.
+//!
+class TableGridBase : public wxGrid {
+  //!
+  //! @name Suppressing the built-in selection highlight & cursor rectangle
+  //!
+  //@{
+  void OnRangeSelecting(wxGridRangeSelectEvent &p_event) { p_event.Veto(); }
+  void OnSelectCell(wxGridEvent &p_event) { p_event.Skip(); }
+  //@}
+
+protected:
+  /// Shows the editor on one click; overridden by the header grids to edit strategies
+  /// instead (or do nothing, if the strategic form is read-only).
+  virtual void OnCellLeftClick(wxGridEvent &p_event)
+  {
+    SetGridCursor(p_event.GetRow(), p_event.GetCol());
+    EnableCellEditControl();
+    p_event.Skip(false);
+  }
+
+public:
+  explicit TableGridBase(wxWindow *p_parent, wxWindowID p_id = wxID_ANY) : wxGrid(p_parent, p_id)
+  {
+    SetCellHighlightPenWidth(0);
+    SetCellHighlightROPenWidth(0);
+    SetRowLabelSize(0);
+    SetColLabelSize(0);
+
+    Bind(wxEVT_GRID_RANGE_SELECTING, &TableGridBase::OnRangeSelecting, this);
+    Bind(wxEVT_GRID_SELECT_CELL, &TableGridBase::OnSelectCell, this);
+    Bind(wxEVT_GRID_CELL_LEFT_CLICK, &TableGridBase::OnCellLeftClick, this);
+  }
+};
 
 class StrategicTableLayout {
   GameDocument *m_doc;
@@ -273,6 +317,22 @@ class TableWidget final : public wxPanel {
   void UpdatePayoffPanel();
   void UpdateLabelPanelMargins();
   void UpdateLabelPanels();
+  //@}
+
+  /// @name Constructing/updating the row/col/payoff grids
+  //@{
+  /// Constructs m_payoffGrid (defined in nfgpayoffgrid.cc, alongside PayoffGrid) and
+  /// m_rowGrid/m_colGrid (defined in nfgheadergrid.cc, alongside RowPlayerGrid/
+  /// ColPlayerGrid) -- called from the constructor body rather than its initializer
+  /// list, since those grid classes' complete types aren't visible here.
+  void InitPayoffGrid();
+  void InitHeaderGrids();
+  /// These call down to PayoffGrid::OnUpdate()/RowPlayerGrid::OnUpdate()/
+  /// ColPlayerGrid::OnUpdate() (via a dynamic_cast back to the concrete type), for the
+  /// same reason InitPayoffGrid()/InitHeaderGrids() above exist.
+  void UpdatePayoffGrid();
+  void UpdateRowGrid();
+  void UpdateColGrid();
   //@}
 
   int GetRowPaneWidth() const;
